Fix #10684: Voice commands were not working in UWP
As we delayed its registration due to .NET 6 problems in Android, it was not ready on time for MRTK scene managers. We move it back to XrvService constructor, as .NET6 issue finally was more related with registering thins before Application.Initialize() invoke. Reverting this part seems not to break .NET6 app.